criado campo generico para DRE
